+++ pm/CHANGELOG.md
@@ -20,6 +20,7 @@ description: A detailed, chronological record of all changes made to a project, 
 - Improved translation context awareness
 - Fixed username matching on the "Character deletion" modal
 - Allowed to add contact of deleted characters
+- Allowed view chat history for deleted characters
 
 ## [0.1.0-7] - 2025-04-01